Time: 0 ms (100%), Space: 63 MB (10.98%) - LeetHub

+++ 0011-container-with-most-water/0011-container-with-most-water.cpp
@@ -4,10 +4,10 @@ public:
         int n=height.size();
         int i=0;
         int j=n-1;
-        int maxi=0;
+        long long maxi=0;
         while(i<j){
             int minheight=min(height[i],height[j]);
-            int area=minheight*(j-i);
+            long long area=minheight*(j-i);
             maxi=max(maxi,area);
             if(height[i]<height[j]) i++;
             else j--;
